A redesign of the "dash" mechanisms for the Titan.
Essentially, I did not like the thematic connection of the Titan providing block and momentum as those should be reserved for your pilot. Throw makes sense to give you momentum as the idea is that your Titan is throwing your titan.

Now, "Dashing" has to do with manipulating  what titan cards you have available. This is also just a card archetype I've been wanting, so this feels like a nice inclusion.

Finally, phase dash was a little too powerful, so I've made it an "engage" card so that it messes with Advantage.