This card crashes my game, and I can't figure out why.
It's a common 0 cost card that draws you cards and places Dazed status' ontop of your deck. We currently only have 2 cards that give positive card draw, which Advantage seriously needs, and those are uncommon and rare.

Anyways, I copied the images for Defend in both the originals folder and the generated folder (in the generated folder, I copied both necessary images).

I'm not sure why the card crashes my game when I tried to play it.